Atualizando versão dos pacotes
