+++ src/Optsol.Components.Domain/Optsol.Components.Domain.csproj
@@ -4,7 +4,7 @@
          <Version>2.3.0</Version>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="FluentValidation" Version="10.3.6" />
+    <PackageReference Include="FluentValidation" Version="11.4.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Optsol.Components.Shared\Optsol.Components.Shared.csproj" />
