+++ src/Optsol.Components.Infra.ElasticSearch/Optsol.Components.Infra.ElasticSearch.csproj
@@ -4,7 +4,7 @@
     <Version>2.3.0</Version>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="NEST" Version="7.17.0" />
+    <PackageReference Include="NEST" Version="7.17.5" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Optsol.Components.Infra\Optsol.Components.Infra.csproj" />
