+++ src/Optsol.Components.Infra.Security.AzureB2C/Security/Extensions/ConfigurationBuilderExtensions.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.Configuration;
+using Optsol.Components.Shared.Extensions;
+
+namespace Microsoft.Extensions.DependencyInjection;
+
+public static class ConfigurationBuilderExtensions
+{
+    public static void AddInMemoryObject(this ConfigurationBuilder configurationBuilder, object settings, string settingsRoot)
+    {
+        configurationBuilder.AddInMemoryCollection(settings.ToKeyValuePairs(settingsRoot));
+    }       
+}
