+++ src/Optsol.Components.Shared/Settings/CorsSettings.cs
@@ -1,24 +1,32 @@
-﻿using System;
+﻿#nullable enable
+using System;
 using System.Collections.Generic;
 
 namespace Optsol.Components.Shared.Settings
 {
     public class CorsSettings : BaseSettings
     {
-        public CorsPolicy DefaultPolicy { get; set; }
-        public IEnumerable<CorsPolicy> Policies { get; set; }
+        public CorsPolicy? DefaultPolicy { get; set; }
+        public IEnumerable<CorsPolicy>? Policies { get; set; }
 
         public override void Validate()
         {
-            DefaultPolicy.Validate();
+            DefaultPolicy?.Validate();
+
+            if (Policies is null) return;
+            
+            foreach (var policy in Policies)
+            {
+                policy.Validate();
+            }
         }
     }
 
     public class CorsPolicy : BaseSettings
     {
-        public string Name { get; set; }
+        public string? Name { get; set; }
 
-        public string[] Origins { get; set; } = Array.Empty<string>();
+        public string[] Origins { get; init; } = Array.Empty<string>();
 
         public override void Validate()
         {
